spark/pyspark_codebase/pyproject.toml
[project]
name = "acme-data-lakehouse"
version = "1.2.0"
description = "ACME Corp Data Lakehouse Pipeline"
requires-python = ">=3.9"

dependencies = [
    "pyspark>=3.5.0",
    "delta-spark>=3.0.0",
    "pyyaml>=6.0",
    "great-expectations>=0.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "chispa>=0.9.0",
]
